requirements.txt
psutil==5.9.5
pymodbus==3.3.0
pyserial==3.5
uModbus==1.0.4
